change route upload file out of project, log debug spring, decorate,fix management time page,add reflection java validation.

See merge request shsoft/mini-erp-team2!171

+++ src/main/java/com/shsoftvina/erpshsoftvina/service/impl/ContractServiceImpl.java
@@ -2,7 +2,7 @@ package com.shsoftvina.erpshsoftvina.service.impl;
 
 import com.shsoftvina.erpshsoftvina.constant.ContractConstant;
 import com.shsoftvina.erpshsoftvina.converter.ContractConverter;
-import com.shsoftvina.erpshsoftvina.entity.Allowance;
+import com.shsoftvina.erpshsoftvina.entity.AllowanceInsurance;
 import com.shsoftvina.erpshsoftvina.entity.Contract;
 import com.shsoftvina.erpshsoftvina.enums.contract.StatusContractEnum;
 import com.shsoftvina.erpshsoftvina.exception.NotFoundException;
@@ -11,12 +11,11 @@ import com.shsoftvina.erpshsoftvina.mapper.UserMapper;
 import com.shsoftvina.erpshsoftvina.model.request.contract.CreateContractRequest;
 import com.shsoftvina.erpshsoftvina.model.request.contract.UpdateContractRequest;
 import com.shsoftvina.erpshsoftvina.model.response.contract.ContractResponse;
-import com.shsoftvina.erpshsoftvina.service.AllowanceService;
+import com.shsoftvina.erpshsoftvina.service.AllowanceInsuranceService;
 import com.shsoftvina.erpshsoftvina.service.ContractService;
 import com.shsoftvina.erpshsoftvina.utils.ApplicationUtils;
 import com.shsoftvina.erpshsoftvina.utils.FileUtils;
 import com.shsoftvina.erpshsoftvina.utils.MessageErrorUtils;
-import liquibase.pro.packaged.E;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -43,7 +42,7 @@ public class ContractServiceImpl implements ContractService {
     private ApplicationUtils applicationUtils;
 
     @Autowired
-    private AllowanceService allowanceService;
+    private AllowanceInsuranceService allowanceInsuranceService;
 
     @Override
     public ContractResponse addContract(CreateContractRequest createContractRequest) {
@@ -71,9 +70,9 @@ public class ContractServiceImpl implements ContractService {
             try {
                 contractMapper.addContract(c);
 
-                List<Allowance> allowances = allowanceService.insertAllowances(c.getId(), createContractRequest.getAllowance());
-                if(allowances != null){
-                    c.setAllowances(allowances);
+                List<AllowanceInsurance> allowanceInsurances = allowanceInsuranceService.insertAllowanceInsurances(c.getId(), createContractRequest.getAllowanceInsurances());
+                if(allowanceInsurances != null){
+                    c.setAllowanceInsurances(allowanceInsurances);
                     return contractConverter.toResponse(c);
                 } else{
                     contractMapper.deleteById(c.getId());
@@ -95,23 +94,17 @@ public class ContractServiceImpl implements ContractService {
     public int updateContract(UpdateContractRequest updateContractRequest) {
 
         Contract contract = contractMapper.findById(updateContractRequest.getId());
-
         if(contract == null) throw new NotFoundException("id");
-
-        int deleteAllowances = allowanceService.deleteAllowances(contract.getId());
-        if (deleteAllowances == 0) throw new NotFoundException("DELETE ALLOWANCE FAIL");
-
-        MultipartFile contractFile = updateContractRequest.getContract();
-
-        if(contractFile!=null) applicationUtils.checkValidateFile(Contract.class, contractFile);
 
         String fileNameContract = null;
         boolean isSaveContractSuccess;
 
+        MultipartFile contractFile = updateContractRequest.getFile();
+        if(contractFile!=null) applicationUtils.checkValidateFile(Contract.class, contractFile);
         if(contractFile != null){
             fileNameContract = FileUtils.formatNameImage(contractFile);
             isSaveContractSuccess = FileUtils.saveImageToServer(
-                    request, ContractConstant.UPLOAD_FILE_DIR, updateContractRequest.getContract(), fileNameContract);
+                    request, ContractConstant.UPLOAD_FILE_DIR, updateContractRequest.getFile(), fileNameContract);
         } else isSaveContractSuccess = false;
 
         Contract c;
@@ -119,13 +112,15 @@ public class ContractServiceImpl implements ContractService {
             c = contractConverter.toEntity(updateContractRequest, fileNameContract);
             try {
                 contractMapper.updateContract(c);
-                List<Allowance> allowances = allowanceService.insertAllowances(c.getId(), updateContractRequest.getAllowance());
-                if(allowances == null) {
+                List<AllowanceInsurance> allowanceInsurances = allowanceInsuranceService.updateAllowanceInsurances(c.getId(), updateContractRequest.getAllowanceInsurances());
+                if(allowanceInsurances == null) {
                     contractMapper.deleteById(c.getId());
                     FileUtils.deleteImageFromServer(request, ContractConstant.UPLOAD_FILE_DIR, fileNameContract);
+                    return 0;
+                } else{
+                    FileUtils.deleteImageFromServer(request, ContractConstant.UPLOAD_FILE_DIR, contract.getFiles());
+                    return 1;
                 }
-                FileUtils.deleteImageFromServer(request, ContractConstant.UPLOAD_FILE_DIR, contract.getFiles());
-                return 1;
             } catch (Exception e){
                 FileUtils.deleteImageFromServer(request, ContractConstant.UPLOAD_FILE_DIR, fileNameContract);
                 return 0;
@@ -135,13 +130,13 @@ public class ContractServiceImpl implements ContractService {
             c = contractConverter.toEntity(updateContractRequest, fileNameContract);
             try{
                 contractMapper.updateContract(c);
-                List<Allowance> allowances = allowanceService.insertAllowances(c.getId(), updateContractRequest.getAllowance());
-                if(allowances == null) {
+                List<AllowanceInsurance> allowanceInsurances = allowanceInsuranceService.updateAllowanceInsurances(c.getId(), updateContractRequest.getAllowanceInsurances());
+                if(allowanceInsurances == null) {
                     contractMapper.deleteById(c.getId());
-                    FileUtils.deleteImageFromServer(request, ContractConstant.UPLOAD_FILE_DIR, fileNameContract);
+                    return 0;
                 }
             }catch (Exception e) {
-                System.out.println(e);
+                return 0;
             }
             return 1;
         }
